Create a rover server that will answer to the rover control center
This kata may be done either with groups doing the rover code and the
control center code or with only groups doing the control center code.
In the later case, we need a fake robot to allow them to run their
progam against. This fake RoverServer will be able to introduce latency,
random failures, etc.